+++ src/JsRoot/GraphData.tsx
@@ -4,7 +4,7 @@ import JsRootGraph2D from './JsRootGraph2D';
 import JsRootGraph0D from './JsRootGraph0D';
 import { Button, Grid } from '@mui/material';
 import { saveString } from '../util/File';
-import { estimatorPageToCsv } from '../util/Csv';
+import { estimatorPageToCsv } from '../util/csv/Csv';
 
 export type pageData = {
 	name: string;
